Load content from a MetaboLights study

The MetaboLightsParam class and the associated readMsObject() method allow users to load an MsExperiment object from a study in the MetaboLights database (https://www.ebi.ac.uk/metabolights/index) by providing its unique study studyId. This function is particularly useful for importing metabolomics data into an MsExperiment object for further analysis within the R environment. It's important to note that this method can only be used for import into an R environement using readMsObject(). It cannot be used with the saveMsObject() method.

If the study contains multiple assays, the user will be prompted to select which assay to load. The resulting MsExperiment object will include a sampleData slot populated with data extracted from the selected assay. Columns in the sampleData that contain only NA values are automatically removed, and an additional column is added to track the injection index.

Usage

MetaboLightsParam(studyId = character(1))

# S4 method for class 'MsExperiment,MetaboLightsParam'
readMsObject(object, param, ...)

Arguments

studyId

character(1) The MetaboLights study studyId, which should start with "MTBL". This identifier uniquely specifies the study within the MetaboLights database.

object

for saveMsObject(): the MS data object to save, for readMsObject(): the MS data object that should be returned

param

an object defining and (eventually configuring) the file format and file name or directory to/from which the data object should be exported/imported.

...

additional optional arguments. See documentation of respective method for more information.

Value

(for now ?) A MsExperiment object with only the sampleData slots filled (will be updated when MetaboLightsBackend available ?).

See also

Author

Philippine Louail

Examples

library(MsExperiment)
# Load a study with the studyId "MTBLS10035"
param <- MetaboLightsParam(studyId = "MTBLS10035")
ms_experiment <- readMsObject(MsExperiment(), param)
#> Only one assay file found: a_MTBLS10035_LC-MS_positive_reverse-phase_metabolite_profiling.txt